+++ test/test_s1_orbits.py
@@ -12,34 +12,63 @@ def test_ensure_orbit_credentials(monkeypatch):
         def __init__(self, netrc_file):
             self.netrc_file = netrc_file
             self.hosts = {}
+
         def __str__(self):
             return str(self.hosts)
 
-    # No .netrc, no ESA CDSE env variables
+    # No .netrc, no ESA CDSE or Earthdata env variables
     with monkeypatch.context() as mp:
         mp.setattr(netrc, 'netrc', EmptyNetrc, raising=False)
         mp.delenv('ESA_USERNAME', raising=False)
         mp.delenv('ESA_PASSWORD', raising=False)
+        mp.delenv('EARTHDATA_USERNAME', raising=False)
+        mp.delenv('EARTHDATA_PASSWORD', raising=False)
         with pytest.raises(ValueError):
             s1_orbits.ensure_orbit_credentials()
 
-    # No .netrc, set ESA CDSE env variables
+    # No .netrc or Earthdata env vars, set ESA CDSE env variables
     with monkeypatch.context() as mp:
         mp.setattr(netrc, 'netrc', EmptyNetrc, raising=False)
         mp.setenv('ESA_USERNAME', 'foo')
         mp.setenv('ESA_PASSWORD', 'bar')
+        mp.delenv('EARTHDATA_USERNAME', raising=False)
+        mp.delenv('EARTHDATA_PASSWORD', raising=False)
+        with pytest.raises(ValueError):
+            s1_orbits.ensure_orbit_credentials()
+
+    # No .netrc or ESA CDSE env vars, set Earthdata env variables
+    with monkeypatch.context() as mp:
+        mp.setattr(netrc, 'netrc', EmptyNetrc, raising=False)
+        mp.delenv('ESA_USERNAME', raising=False)
+        mp.delenv('ESA_PASSWORD', raising=False)
+        mp.setenv('EARTHDATA_USERNAME', 'fizz')
+        mp.setenv('EARTHDATA_PASSWORD', 'buzz')
+        with pytest.raises(ValueError):
+            s1_orbits.ensure_orbit_credentials()
+
+    # No .netrc, set Earthdata and ESA CDSE env variables
+    with monkeypatch.context() as mp:
+        mp.setattr(netrc, 'netrc', EmptyNetrc, raising=False)
+        mp.setenv('ESA_USERNAME', 'foo')
+        mp.setenv('ESA_PASSWORD', 'bar')
+        mp.setenv('EARTHDATA_USERNAME', 'fizz')
+        mp.setenv('EARTHDATA_PASSWORD', 'buzz')
         mp.setattr(Path, 'write_text', lambda self, write_text: write_text)
         written_credentials = s1_orbits.ensure_orbit_credentials()
-        assert written_credentials == str({s1_orbits.ESA_CDSE_HOST: ('foo', None, 'bar')})
+        assert written_credentials == str({
+            s1_orbits.ESA_CDSE_HOST: ('foo', None, 'bar'),
+            s1_orbits.NASA_EDL_HOST: ('fizz', None, 'buzz')
+        })
 
     class NoCDSENetrc():
         def __init__(self, netrc_file):
             self.netrc_file = netrc_file
-            self.hosts = {'fizz.buzz.org': ('foo', None, 'bar')}
+            self.hosts = {s1_orbits.NASA_EDL_HOST: ('fizz', None, 'buzz')}
+
         def __str__(self):
             return str(self.hosts)
 
-    # No CDSE in .netrc, no ESA CDSE env variables
+    # No CDSE in .netrc or ESA CDSE env variables, Earthdata in .netrc
     with monkeypatch.context() as mp:
         mp.setattr(netrc, 'netrc', NoCDSENetrc, raising=False)
         mp.delenv('ESA_USERNAME', raising=False)
@@ -47,35 +76,81 @@ def test_ensure_orbit_credentials(monkeypatch):
         with pytest.raises(ValueError):
             s1_orbits.ensure_orbit_credentials()
 
-    # No CDSE in .netrc, set ESA CDSE env variables
+    # No CDSE in .netrc, set ESA CDSE env variables, Earthdata in .netrc
     with monkeypatch.context() as mp:
         mp.setattr(netrc, 'netrc', NoCDSENetrc, raising=False)
         mp.setenv('ESA_USERNAME', 'foo')
         mp.setenv('ESA_PASSWORD', 'bar')
         mp.setattr(Path, 'write_text', lambda self, write_text: write_text)
         written_credentials = s1_orbits.ensure_orbit_credentials()
-        assert written_credentials == str({'fizz.buzz.org': ('foo', None, 'bar'), s1_orbits.ESA_CDSE_HOST: ('foo', None, 'bar')})
+        assert written_credentials == str({
+            s1_orbits.NASA_EDL_HOST: ('fizz', None, 'buzz'),
+            s1_orbits.ESA_CDSE_HOST: ('foo', None, 'bar'),
+        })
 
-    class CDSENetrc():
+    class NoEarthdataNetrc():
         def __init__(self, netrc_file):
             self.netrc_file = netrc_file
             self.hosts = {s1_orbits.ESA_CDSE_HOST: ('foo', None, 'bar')}
+
         def __str__(self):
             return str(self.hosts)
 
-    # cdse in .netrc, no ESA CDSE env variables
+    # cdse in .netrc, no ESA CDSE env variables, Earthdata env variables
     with monkeypatch.context() as mp:
-        mp.setattr(netrc, 'netrc', CDSENetrc, raising=False)
+        mp.setattr(netrc, 'netrc', NoEarthdataNetrc, raising=False)
         mp.delenv('ESA_USERNAME', raising=False)
         mp.delenv('ESA_PASSWORD', raising=False)
+        mp.setenv('EARTHDATA_USERNAME', 'fizz')
+        mp.setenv('EARTHDATA_PASSWORD', 'buzz')
+        mp.setattr(Path, 'write_text', lambda self, write_text: write_text)
+        written_credentials = s1_orbits.ensure_orbit_credentials()
+        assert written_credentials == str({
+            s1_orbits.ESA_CDSE_HOST: ('foo', None, 'bar'),
+            s1_orbits.NASA_EDL_HOST: ('fizz', None, 'buzz'),
+        })
+
+    class CDSEAndEarthdataNetrc():
+        def __init__(self, netrc_file):
+            self.netrc_file = netrc_file
+            self.hosts = {
+                s1_orbits.ESA_CDSE_HOST: ('foo', None, 'bar'),
+                s1_orbits.NASA_EDL_HOST: ('fizz', None, 'buzz')
+            }
+
+        def __str__(self):
+            return str(self.hosts)
+
+    # cdse and Earthdata in netrc, no env variables
+    with monkeypatch.context() as mp:
+        mp.setattr(netrc, 'netrc', CDSEAndEarthdataNetrc, raising=False)
         written_credentials = s1_orbits.ensure_orbit_credentials()
         assert written_credentials is None
 
-    # cdse in .netrc, set ESA CDSE env variables
     with monkeypatch.context() as mp:
-        mp.setattr(netrc, 'netrc', CDSENetrc, raising=False)
+        mp.setattr(netrc, 'netrc', CDSEAndEarthdataNetrc, raising=False)
+        mp.delenv('ESA_USERNAME', raising=False)
+        mp.delenv('ESA_PASSWORD', raising=False)
+        mp.setenv('EARTHDATA_USERNAME', 'fizz')
+        mp.setenv('EARTHDATA_PASSWORD', 'buzz')
+        written_credentials = s1_orbits.ensure_orbit_credentials()
+        assert written_credentials is None
+
+    with monkeypatch.context() as mp:
+        mp.setattr(netrc, 'netrc', CDSEAndEarthdataNetrc, raising=False)
         mp.setenv('ESA_USERNAME', 'foo')
         mp.setenv('ESA_PASSWORD', 'bar')
+        mp.setenv('EARTHDATA_USERNAME', 'fizz')
+        mp.setenv('EARTHDATA_PASSWORD', 'buzz')
+        written_credentials = s1_orbits.ensure_orbit_credentials()
+        assert written_credentials is None
+
+    with monkeypatch.context() as mp:
+        mp.setattr(netrc, 'netrc', CDSEAndEarthdataNetrc, raising=False)
+        mp.setenv('ESA_USERNAME', 'foo')
+        mp.setenv('ESA_PASSWORD', 'bar')
+        mp.delenv('ESA_USERNAME', raising=False)
+        mp.delenv('ESA_PASSWORD', raising=False)
         written_credentials = s1_orbits.ensure_orbit_credentials()
         assert written_credentials is None