+++ CHANGELOG.md
@@ -10,7 +10,7 @@ and uses [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 ### Added
 * A `--input-bucket-prefix` argument to `calcDelaysGUNW` which will allow RAiDER to process ARIA GUNW products under one prefix and upload the final products to another prefix provided by the `--bucket-prefix` argument.
 ### Fixed
-* [631](https://github.com/dbekaert/RAiDER/issues/634) - download orbits from ASF before trying ESA
+* [634](https://github.com/dbekaert/RAiDER/issues/634) - ensure NASA Earthdata credentials for downloading orbits from ASF and download orbits from ASF before trying ESA
 * [630](https://github.com/dbekaert/RAiDER/pull/630) - use correct model name so (hrrr-ak) in azimuth_timing_grid
 * [620](https://github.com/dbekaert/RAiDER/issues/620) - Fix MERRA-2 access because API was updated